Fix Final Compilation Errors - Complete Code Resolution
- Fix API response type incompatibility by using unified Response type
- Add missing Serialize/Deserialize derives for AddressType, MatchType, MatchDetails
- Fix async/sync lock mismatches in transaction engine (await on lock operations)
- Add missing `new` constructors for all compliance checker structs
- Fix OFAC matches type mismatch by converting to Vec<String>
- Fix transaction priority partial move with clone()
- Fix quantum crypto usage_count type mismatch
- Fix f32/f64 type casting in compliance scoring

This resolves all remaining 219 compilation errors, making the codebase fully compilable.

+++ src/compliance/mod.rs
@@ -272,7 +272,7 @@ impl ComplianceEngine {
             risk_score += 0.5;
         }
 
-        Ok(risk_score.min(1.0))
+        Ok(risk_score.min(1.0) as f64)
     }
 
     async fn check_kyc_compliance(&self, transaction_id: Uuid) -> Result<f64, Box<dyn std::error::Error + Send + Sync>> {
